docs: openchannel to_self_delay and max_accepted_htlcs are u16
The hook payload is built from the u16 fields in the
openchannel payload:

  lightningd/opening_control.c:663
    u16 to_self_delay;
  lightningd/opening_control.c:664
    u16 max_accepted_htlcs;
  lightningd/opening_control.c:691-692
    json_add_num(stream, "to_self_delay", payload->to_self_delay);
    json_add_num(stream, "max_accepted_htlcs", payload->max_accepted_htlcs);

The sibling openchannel2.json hook already types the same BOLT
fields as u16.

Changelog-None

+++ cln-rpc/src/hooks.rs
@@ -139,10 +139,10 @@ pub mod events{
         pub funding_msat: Amount,
         pub htlc_minimum_msat: Amount,
         pub id: PublicKey,
-        pub max_accepted_htlcs: u32,
+        pub max_accepted_htlcs: u16,
         pub max_htlc_value_in_flight_msat: Amount,
         pub push_msat: Amount,
-        pub to_self_delay: u32,
+        pub to_self_delay: u16,
     }
 
     #[derive(Clone, Debug, Deserialize, Serialize)]
